+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 12: Tracalorie Project (Using The Module Pattern)
     - Add Total Calories
+    - Working With The Edit State
 - Logic to update and display total calories was added.
+- Edit state base logic was added.
 
 ## October 6, 2020
 - Modern JavaScript From The Beginning (Udemy)